No tmp

+++ run.py
@@ -190,58 +190,6 @@ def main(gtk_context):
         log.info("Using provided PyBIDS filter file %s", str(paths[0]))
         config["bids-filter-file"] = str(paths[0])
 
-    previous_work_zip_file_path = gtk_context.get_input_path("work-dir")
-    if previous_work_zip_file_path:
-        paths = list(Path("input/work-dir").glob("*"))
-        log.info("Using provided fMRIPrep intermediate work file %s", str(paths[0]))
-        unzip_dir = FWV0 / "unzip-work-dir"
-        unzip_dir.mkdir(parents=True)
-        unzip_archive(paths[0], unzip_dir)
-        for a_dir in unzip_dir.glob("*/*"):
-            if (
-                a_dir.name == "bids"
-            ):  # skip previous bids directory so current bids data will be used
-                log.info(
-                    "Found %s, but ignoring it to use current bids data", a_dir.name
-                )
-            else:
-                log.info("Found %s", a_dir.name)
-                a_dir.rename(FWV0 / "work" / a_dir.name)
-        hash_file = list(Path("work/fmriprep_wf/").glob("fsdir_run_*/_0x*.json"))[0]
-        if hash_file.exists():
-            with open(hash_file) as json_file:
-                data = json.load(json_file)
-                old_tmp_path = data[0][1]
-                old_tmp_name = old_tmp_path.split("/")[2]
-                log.info("Found old tmp name: %s", old_tmp_name)
-                cur_tmp_name = str(FWV0).split("/")[2]
-                # rename the directory to the old name
-                Path("/tmp/" + cur_tmp_name).replace(Path("/tmp/" + old_tmp_name))
-                # create a symbolic link using the new name to the old name just in case
-                Path("/tmp/" + cur_tmp_name).symlink_to(
-                    Path("/tmp/" + old_tmp_name), target_is_directory=True
-                )
-                # update all variables to have the old directory name in them
-                FWV0 = Path("/tmp/" + old_tmp_name + "/flywheel/v0")
-                output_dir = str(output_dir).replace(cur_tmp_name, old_tmp_name)
-                output_analysis_id_dir = Path(
-                    str(output_analysis_id_dir).replace(cur_tmp_name, old_tmp_name)
-                )
-                log.info("new output directory is: %s", output_dir)
-                work_dir = Path(str(work_dir).replace(cur_tmp_name, old_tmp_name))
-                log.info("new work directory is: %s", work_dir)
-                subjects_dir = Path(
-                    str(subjects_dir).replace(cur_tmp_name, old_tmp_name)
-                )
-                config["fs-subjects-dir"] = subjects_dir
-                log.info("new FreeSurfer subjects directory is: %s", subjects_dir)
-                # for old work to be recognized, switch to running from the old path
-                os.chdir(FWV0)
-                log.info("cd %s", FWV0)
-        else:
-            log.info("Could not find hash file")
-        config["work-dir"] = str(FWV0 / "work")
-
     subject_zip_file_path = gtk_context.get_input_path("fs-subjects-dir")
     if subject_zip_file_path:
         paths = list(Path("input/fs-subjects-dir").glob("*"))
@@ -283,11 +231,17 @@ def main(gtk_context):
         FREESURFER_LICENSE,
     )
 
-    # TemplateFlow seems to be baked in to the container since 2021-10-07 16:25:12 so this is not needed
-    # templateflow_dir = FWV0 / "templateflow"
-    # templateflow_dir.mkdir()
-    # environ["SINGULARITYENV_TEMPLATEFLOW_HOME"] = str(templateflow_dir)
-    # environ["TEMPLATEFLOW_HOME"] = str(templateflow_dir)
+    # TemplateFlow seems to be baked in to the container since 2021-10-07 16:25:12 so this is not needed...actually, it is for now...
+    templateflow_dir = FWV0 / "templateflow"
+    templateflow_dir.mkdir()
+    environ["SINGULARITYENV_TEMPLATEFLOW_HOME"] = str(templateflow_dir)
+    environ["TEMPLATEFLOW_HOME"] = str(templateflow_dir)
+    orig = Path("/home/fmriprep/.cache/templateflow/")
+    # Fill writable templateflow directory with existing templates so they don't have to be downloaded
+    templates = list(orig.glob("*"))
+    for tt in templates:
+        # (templateflow_dir / tt.name).symlink_to(tt)
+        shutil.copytree(tt, templateflow_dir / tt.name)
 
     command = generate_command(
         config, work_dir, output_analysis_id_dir, errors, warnings
@@ -317,140 +271,164 @@ def main(gtk_context):
         log.info("Did not download BIDS because of previous errors")
         print(errors)
 
-    # Don't run if there were errors or if this is a dry run
     return_code = 0
+    num_tries = 0
 
-    try:
-
-        if len(errors) > 0:
-            return_code = 1
-            log.info("Command was NOT run because of previous errors.")
-
-        elif dry_run:
-            e = "gear-dry-run is set: Command was NOT run."
-            log.warning(e)
-            warnings.append(e)
-            pretend_it_ran(destination_id)
-
-        else:
-            if config["gear-log-level"] != "INFO":
-                # show what's in the current working directory just before running
-                os.system("tree -al .")
-
-            if "gear-timeout" in config:
-                command = [f"timeout {config['gear-timeout']}"] + command
-
-            # This is what it is all about
-            exec_command(
-                command, environ=environ, dry_run=dry_run, shell=True, cont_output=True,
-            )
-
-    except RuntimeError as exc:
+    # Don't run if there were errors or if this is a dry run
+    if len(errors) > 0:
         return_code = 1
-        errors.append(exc)
-        log.critical(exc)
-        log.exception("Unable to execute command.")
+        log.info("Command was NOT run because of previous errors.")
+        num_tries == 2  # don't try to run
 
-    finally:
+    while num_tries < 2:
 
-        # Save time, etc. resources used in metadata on analysis
-        if Path("time_output.txt").exists():  # some tests won't have this file
-            metadata = {
-                "analysis": {"info": {"resources used": {},},},
-            }
-            with open("time_output.txt") as file:
-                for line in file:
-                    if ":" in line:
-                        if (
-                            "Elapsed" in line
-                        ):  # special case "Elapsed (wall clock) time (h:mm:ss or m:ss): 0:08.11"
-                            sline = re.split(r"\):", line)
-                            sline[0] += ")"
-                        else:
-                            sline = line.split(":")
-                        key = sline[0].strip()
-                        val = sline[1].strip(' "\n')
-                        metadata["analysis"]["info"]["resources used"][key] = val
-            with open(f"{output_dir}/.metadata.json", "w") as fff:
-                json.dump(metadata, fff)
-                log.info(f"Wrote {output_dir}/.metadata.json")
+        try:
 
-        # Cleanup, move all results to the output directory
+            num_tries += 1
+            if num_tries > 1:
+                log.info("Trying a second time")
 
-        # Remove all fsaverage* directories
-        if not config.get("gear-keep-fsaverage"):
-            path = output_analysis_id_dir / "freesurfer"
-            fsavg_dirs = path.glob("fsaverage*")
-            for fsavg in fsavg_dirs:
-                log.info("deleting %s", str(fsavg))
-                shutil.rmtree(fsavg)
-        else:
-            log.info("Keeping fsaverage directories")
-
-        # zip entire output/<analysis_id> folder into
-        #  <gear_name>_<project|subject|session label>_<analysis.id>.zip
-        zip_file_name = gear_name + f"_{run_label}_{destination_id}.zip"
-        zip_output(
-            str(output_dir),
-            destination_id,
-            zip_file_name,
-            dry_run=False,
-            exclude_files=None,
-        )
-
-        # Make archives for result *.html files for easy display on platform
-        zip_htmls(output_dir, destination_id, output_analysis_id_dir / BIDS_APP)
-
-        # possibly save ALL intermediate output
-        if config.get("gear-save-intermediate-output"):
-            zip_all_intermediate_output(
-                destination_id, gear_name, output_dir, work_dir, run_label
-            )
-
-        # possibly save intermediate files and folders
-        zip_intermediate_selected(
-            config.get("gear-intermediate-files"),
-            config.get("gear-intermediate-folders"),
-            destination_id,
-            gear_name,
-            output_dir,
-            work_dir,
-            run_label,
-        )
-
-        # clean up: remove output that was zipped
-        if Path(output_analysis_id_dir).exists():
-            if not config.get("gear-keep-output"):
-
-                log.debug('removing output directory "%s"', str(output_analysis_id_dir))
-                shutil.rmtree(output_analysis_id_dir)
+            if dry_run:
+                e = "gear-dry-run is set: Command was NOT run."
+                log.warning(e)
+                warnings.append(e)
+                pretend_it_ran(destination_id)
 
             else:
-                log.info(
-                    'NOT removing output directory "%s"', str(output_analysis_id_dir)
+
+                if "gear-timeout" in config:
+                    command = [f"timeout {config['gear-timeout']}"] + command
+
+                if config["gear-log-level"] != "INFO":
+                    # show what's in the current working directory just before running
+                    os.system("tree -alh .")
+
+                # This is what it is all about
+                exec_command(
+                    command,
+                    environ=environ,
+                    dry_run=dry_run,
+                    shell=True,
+                    cont_output=True,
                 )
 
-        else:
-            log.info("Output directory does not exist so it cannot be removed")
-
-        # Report errors and warnings at the end of the log so they can be easily seen.
-        if len(warnings) > 0:
-            msg = "Previous warnings:\n"
-            for warn in warnings:
-                msg += "  Warning: " + str(warn) + "\n"
-            log.info(msg)
-
-        if len(errors) > 0:
-            msg = "Previous errors:\n"
-            for err in errors:
-                if str(type(err)).split("'")[1] == "str":
-                    # show string
-                    msg += "  Error msg: " + str(err) + "\n"
-                else:  # show type (of error) and error message
-                    err_type = str(type(err)).split("'")[1]
-                    msg += f"  {err_type}: {str(err)}\n"
-            log.info(msg)
+        except RuntimeError as exc:
             return_code = 1
+            errors.append(exc)
+            log.critical(exc)
+            log.exception("Unable to execute command.")
+
+    # Save time, etc. resources used in metadata on analysis
+    if Path("time_output.txt").exists():  # some tests won't have this file
+        metadata = {
+            "analysis": {"info": {"resources used": {},},},
+        }
+        with open("time_output.txt") as file:
+            for line in file:
+                if ":" in line:
+                    if (
+                        "Elapsed" in line
+                    ):  # special case "Elapsed (wall clock) time (h:mm:ss or m:ss): 0:08.11"
+                        sline = re.split(r"\):", line)
+                        sline[0] += ")"
+                    else:
+                        sline = line.split(":")
+                    key = sline[0].strip()
+                    val = sline[1].strip(' "\n')
+                    metadata["analysis"]["info"]["resources used"][key] = val
+        with open(f"{output_dir}/.metadata.json", "w") as fff:
+            json.dump(metadata, fff)
+            log.info(f"Wrote {output_dir}/.metadata.json")
+
+    # Cleanup, move all results to the output directory
+
+    if return_code != 0:
+        os.system("echo ")
+        os.system("echo Disk Information on Failure")
+        os.system("df -h")
+
+    # Remove all fsaverage* directories
+    if not config.get("gear-keep-fsaverage"):
+        path = output_analysis_id_dir / "freesurfer"
+        fsavg_dirs = path.glob("fsaverage*")
+        for fsavg in fsavg_dirs:
+            log.info("deleting %s", str(fsavg))
+            shutil.rmtree(fsavg)
+    else:
+        log.info("Keeping fsaverage directories")
+
+    # zip entire output/<analysis_id> folder into
+    #  <gear_name>_<project|subject|session label>_<analysis.id>.zip
+    zip_file_name = gear_name + f"_{run_label}_{destination_id}.zip"
+    zip_output(
+        str(output_dir),
+        destination_id,
+        zip_file_name,
+        dry_run=False,
+        exclude_files=None,
+    )
+
+    # Make archives for result *.html files for easy display on platform
+    zip_htmls(output_dir, destination_id, output_analysis_id_dir / BIDS_APP)
+
+    # possibly save ALL intermediate output
+    if config.get("gear-save-intermediate-output"):
+        zip_all_intermediate_output(
+            destination_id, gear_name, output_dir, work_dir, run_label
+        )
+
+    # possibly save intermediate files and folders
+    zip_intermediate_selected(
+        config.get("gear-intermediate-files"),
+        config.get("gear-intermediate-folders"),
+        destination_id,
+        gear_name,
+        output_dir,
+        work_dir,
+        run_label,
+    )
+
+    # clean up: remove output that was zipped
+    if Path(output_analysis_id_dir).exists():
+        if not config.get("gear-keep-output"):
+
+            log.debug('removing output directory "%s"', str(output_analysis_id_dir))
+            shutil.rmtree(output_analysis_id_dir)
+
+        else:
+            log.info('NOT removing output directory "%s"', str(output_analysis_id_dir))
+
+    else:
+        log.info("Output directory does not exist so it cannot be removed")
+
+    # Report errors and warnings at the end of the log so they can be easily seen.
+    if len(warnings) > 0:
+        msg = "Previous warnings:\n"
+        for warn in warnings:
+            msg += "  Warning: " + str(warn) + "\n"
+        log.info(msg)
+
+    if len(errors) > 0:
+        msg = "Previous errors:\n"
+        for err in errors:
+            if str(type(err)).split("'")[1] == "str":
+                # show string
+                msg += "  Error msg: " + str(err) + "\n"
+            else:  # show type (of error) and error message
+                err_type = str(type(err)).split("'")[1]
+                msg += f"  {err_type}: {str(err)}\n"
+        log.info(msg)
+        return_code = 1
+
+    if num_tries == 1:
+        log.info("Happily, fMRIPrep worked on the first try.")
+    else:
+        msg = (
+            "first try but it did on the second"
+            if return_code == 0
+            else "first or second try"
+        )
+        log.info("Sadly, fMRIPrep did not work on the %s.", msg)
 
     log.info("%s Gear is done.  Returning %s", CONTAINER, return_code)
 
@@ -459,18 +437,21 @@ def main(gtk_context):
 
 if __name__ == "__main__":
 
-    # always run in a newly created "scratch" directory in /tmp/...
-    scratch_dir = run_in_tmp_dir()
-
     with flywheel_gear_toolkit.GearToolkitContext() as gtk_context:
+
+        # make sure /flywheel/v0 is writable, use a scratch directory if not
+        scratch_dir = run_in_tmp_dir(gtk_context.config["gear-writable-dir"])
+
         return_code = main(gtk_context)
 
-    # clean up (might be necessary when running in a shared computing environment)
-    for thing in scratch_dir.glob("*"):
-        if thing.is_symlink():
-            thing.unlink()  # don't remove anything links point to
-            log.debug("unlinked %s", thing.name)
-    shutil.rmtree(scratch_dir)
-    log.debug("Removed %s", scratch_dir)
+        # clean up (might be necessary when running in a shared computing environment)
+        if scratch_dir:
+            log.debug("Removing scratch directory")
+            for thing in scratch_dir.glob("*"):
+                if thing.is_symlink():
+                    thing.unlink()  # don't remove anything links point to
+                    log.debug("unlinked %s", thing.name)
+            shutil.rmtree(scratch_dir)
+            log.debug("Removed %s", scratch_dir)
 
     sys.exit(return_code)
